Code Review

+++ Falafel2014PRIMO4586TeamA/src/edu/wpi/first/wpilibj/templates/commands/lifter/MoveToCycles.java
@@ -51,7 +51,7 @@ public class MoveToCycles extends CommandBase {
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
-        return target==SmartDashboard.getNumber("Cycles: ");
+        return target==(int)SmartDashboard.getNumber("Cycles: ");
     }
 
     // Called once after isFinished returns true
